Authorization on parts of the app (#43) [Finishes #170300794]
* :art: Require authentication when performing operations on a Place

* :heavy_plus_sign: Add Pundit into project

* :tada: Initial implementation of PlacePolicy

* :art: Apply PlacePolicy on PlacesController operations

* :wrench: Write helper to use PlacePolicy in the view

* :art: Apply PlacePolicy on elements to be rendered in the view

* :white_check_mark: Update Place system test after applying Pundit

* :white_check_mark: Inlcude conditional rendering of delete link in Place show page

* :sparkles: Apply authorization when creating a new place

+++ app/controllers/places_controller.rb
@@ -1,4 +1,5 @@
 class PlacesController < ApplicationController
+  before_action :require_login, only: %i[new create edit update destroy]
 
   def show
     @place = Place.find(params[:id])
@@ -6,6 +7,7 @@ class PlacesController < ApplicationController
 
   def new
     @user  = User.find(params[:user_id])
+    authorize @user, policy_class: PlacePolicy
     @place = Place.new
   end
 
@@ -25,11 +27,15 @@ class PlacesController < ApplicationController
   def edit
     @place = Place.find(params[:id])
 
+    authorize @place
+
     render :new
   end
 
   def update
     @place = Place.find(params[:id])
+
+    authorize @place
 
     if @place.update(place_params)
       flash[:success] = 'Update success!'
@@ -43,6 +49,8 @@ class PlacesController < ApplicationController
 
   def destroy
     place = Place.find(params[:id])
+
+    authorize place
 
     if place.destroy
       flash[:success] = 'Place deleted successfuly.'
